Factor hdf5-like sources into different files
