+++ gunpowder/nodes/__init__.py
@@ -12,12 +12,13 @@ from .dvid_source import DvidSource
 from .elastic_augment import ElasticAugment
 from .exclude_labels import ExcludeLabels
 from .grow_boundary import GrowBoundary
-from .hdf5_source import Hdf5Source, N5Source, ZarrSource
+from .hdf5_source import Hdf5Source
 from .hdf5_write import Hdf5Write
 from .intensity_augment import IntensityAugment
 from .intensity_scale_shift import IntensityScaleShift
 from .klb_source import KlbSource
 from .merge_provider import MergeProvider
+from .n5_source import N5Source
 from .normalize import Normalize
 from .pad import Pad
 from .precache import PreCache
@@ -31,3 +32,4 @@ from .scan import Scan
 from .simple_augment import SimpleAugment
 from .snapshot import Snapshot
 from .specified_location import SpecifiedLocation
+from .zarr_source import ZarrSource
